removed route handling
