+++ src/utils/findUniqueTotals.js
@@ -17,6 +17,7 @@ export default function findUniqueTotals(callCardData) {
       callsById[call.union_call_id] = {
         instances: [call],
         members_needed: [call.members_needed],
+        uniqueJobsForLifeCycle: []
       };
     }
   });
@@ -30,6 +31,7 @@ export default function findUniqueTotals(callCardData) {
     count ++;
     // travesrse the members_needed array to determine number of real jobs for the lifecycle of the call
     const realJobsCreated = countRealJobs([...members_needed]);
+    callsById[id].uniqueJobsForLifeCycle = realJobsCreated;
 
     if (uniqueJobsByClassification.hasOwnProperty(member_class)) {
       uniqueJobsByClassification[member_class] += realJobsCreated;
